Plotting refactor: Created _plot_commons.py to hold the factorized plotting utilities
